Reorganized screens in different files, started routing
